Fixing failures

+++ app/Http/Controllers/CompanyController.php
@@ -7,8 +7,15 @@ use Illuminate\Http\Request;
 
 class CompanyController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function index()
     {
+        $this->authorize('update');
+
         $companies = Company::all();
 
         return view('admin.companies.index', compact('companies'));
@@ -16,6 +23,8 @@ class CompanyController extends Controller
 
     public function create()
     {
+        $this->authorize('update');
+
         $company = new Company();
 
         return view('admin.companies.create', compact( 'company' ));
@@ -23,6 +32,8 @@ class CompanyController extends Controller
 
     public function store()
     {
+        $this->authorize('update');
+
         Company::create($this->validateRequest());
 
         return redirect('admin/companies');
@@ -30,11 +41,13 @@ class CompanyController extends Controller
 
     public function edit(Company $company)
     {
-        //
+        $this->authorize('update');
     }
 
     public function update(Company $company)
     {
+        $this->authorize('update');
+
         $company->update($this->validateRequest());
 
         return redirect($company->path());
@@ -42,6 +55,8 @@ class CompanyController extends Controller
 
     public function destroy(Company $company)
     {
+        $this->authorize('update');
+
         $company->delete();
 
         return redirect('/admin/companies');
