+++ app/Http/Controllers/RoleController.php
@@ -6,12 +6,19 @@ namespace App\Http\Controllers;
 
 use App\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class RoleController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
 
     public function index()
     {
+        $this->authorize('update');
+
         $roles = Role::all();
 
         return view('admin.roles.index', compact('roles'));
@@ -19,6 +26,8 @@ class RoleController extends Controller
 
     public function create()
     {
+        $this->authorize('update');
+
         $role = new Role();
 
         return view('admin.roles.create', compact( 'role' ));
@@ -26,17 +35,27 @@ class RoleController extends Controller
 
     public function store()
     {
+        $this->authorize('update');
+
         $role = Role::create([
             'name' => request('name'),
             'description' => request('description')
         ]);
 
         return redirect($role->path());
+    }
+
+    public function show()
+    {
+
+        $this->authorize('update');
 
     }
 
     public function update(Role $role)
     {
+        $this->authorize('update');
+
         $role->update($this->validateRequest());
 
         return redirect($role->path());
@@ -44,11 +63,13 @@ class RoleController extends Controller
 
     public function edit(Role $role)
     {
-
+        $this->authorize('update');
     }
 
     public function destroy(Role $role)
     {
+        $this->authorize('update');
+
         $role->delete();
 
         return redirect('/admin/roles');
